Settings consolidation U1: extract diff-option constants to settings/diffOptions.ts
First cut severing the monolith: move the 7 diff-option constant arrays out of Settings.tsx
into a neutral settings/diffOptions.ts; repoint DiffOptionsPopover off the monolith. Also
record the LOCKED decisions (one universal dialog, daemon-aware config store, full migration).
